JAY's FEDEV: styling dialog

+++ lib/providers/time_provider.dart
@@ -4,8 +4,8 @@ final repetitionProvider = StateProvider<double>((ref) {
   return 1.0;
 });
 
-final totalProvider = StateProvider<double>((ref) {
-  return 1.0;
+final totalProvider = StateProvider<int>((ref) {
+  return 1;
 });
 
 final sessionProvider = StateProvider<double>((ref) {
